SASS-2300 created reduced annual amount page. And moved controllers into relevant packages.
