+++ app/controllers/pensions/paymentsIntoPension/RetirementAnnuityAmountController.scala
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package controllers.pensions
+package controllers.pensions.paymentsIntoPension
 
 import config.{AppConfig, ErrorHandler}
-import controllers.pensions.routes._
+import controllers.pensions.paymentsIntoPension.routes._
 import controllers.predicates.AuthorisedAction
-import forms.AmountForm
+import forms.{AmountForm, FormUtils}
 import models.mongo.PensionsCYAModel
 import models.pension.charges.PensionAnnualAllowancesViewModel
 import models.pension.reliefs.PaymentsIntoPensionViewModel
@@ -28,38 +28,40 @@ import play.api.i18n.I18nSupport
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import services.PensionSessionService
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendController
-import utils.Clock
-import views.html.pensions.ReliefAtSourcePaymentsAndTaxReliefAmountView
+import utils.{Clock, SessionHelper}
+import views.html.pensions.RetirementAnnuityAmountView
 
 import javax.inject.{Inject, Singleton}
 import scala.concurrent.Future
 
 
 @Singleton
-class ReliefAtSourcePaymentsAndTaxReliefAmountController @Inject()(implicit val mcc: MessagesControllerComponents,
-                                                                   appConfig: AppConfig,
-                                                                   authAction: AuthorisedAction,
-                                                                   pensionSessionService: PensionSessionService,
-                                                                   errorHandler: ErrorHandler,
-                                                                   view: ReliefAtSourcePaymentsAndTaxReliefAmountView,
-                                                                   clock: Clock) extends FrontendController(mcc) with I18nSupport {
+class RetirementAnnuityAmountController @Inject()(implicit val mcc: MessagesControllerComponents,
+                                                  authAction: AuthorisedAction,
+                                                  retirementAnnuityAmountView: RetirementAnnuityAmountView,
+                                                  appConfig: AppConfig,
+                                                  pensionSessionService: PensionSessionService,
+                                                  errorHandler: ErrorHandler,
+                                                  clock: Clock) extends FrontendController(mcc) with I18nSupport with SessionHelper with FormUtils {
 
   val amountForm: Form[BigDecimal] = AmountForm.amountForm(
-    emptyFieldKey = "pensions.reliefAtSourceTotalPaymentsAndTaxReliefAmount.error.noEntry",
-    wrongFormatKey = "pensions.reliefAtSourceTotalPaymentsAndTaxReliefAmount.error.invalidFormat",
-    exceedsMaxAmountKey = "pensions.reliefAtSourceTotalPaymentsAndTaxReliefAmount.error.overMaximum"
+    emptyFieldKey = "pensions.retirementAnnuityAmount.error.noEntry",
+    wrongFormatKey = "pensions.retirementAnnuityAmount.error.incorrectFormat",
+    exceedsMaxAmountKey = "pensions.retirementAnnuityAmount.error.overMaximum"
   )
+
 
   def show(taxYear: Int): Action[AnyContent] = authAction.async { implicit request =>
     pensionSessionService.getPensionsSessionDataResult(taxYear, request.user) {
       case Some(data) =>
-        if (data.pensions.paymentsIntoPension.rasPensionPaymentQuestion.contains(true)) {
-          data.pensions.paymentsIntoPension.totalRASPaymentsAndTaxRelief match {
-            case Some(amount) => Future.successful(Ok(view(amountForm.fill(amount), taxYear)))
-            case None => Future.successful(Ok(view(amountForm, taxYear)))
+        if (data.pensions.paymentsIntoPension.retirementAnnuityContractPaymentsQuestion.contains(true)) {
+          data.pensions.paymentsIntoPension.totalRetirementAnnuityContractPayments match {
+            case Some(amount) =>
+              Future.successful(Ok(retirementAnnuityAmountView(amountForm.fill(amount), taxYear)))
+            case None => Future.successful(Ok(retirementAnnuityAmountView(amountForm, taxYear)))
           }
         } else {
-          Future.successful(Redirect(ReliefAtSourcePensionsController.show(taxYear)))
+          Future.successful(Redirect(RetirementAnnuityController.show(taxYear)))
         }
       case _ =>
         Future.successful(Redirect(PaymentsIntoPensionsCYAController.show(taxYear)))
@@ -67,24 +69,23 @@ class ReliefAtSourcePaymentsAndTaxReliefAmountController @Inject()(implicit val 
 
   }
 
+
   def submit(taxYear: Int): Action[AnyContent] = authAction.async { implicit request =>
     amountForm.bindFromRequest.fold(
-      formWithErrors => Future.successful(BadRequest(view(formWithErrors, taxYear))),
+      formWithErrors => Future.successful(BadRequest(retirementAnnuityAmountView(formWithErrors, taxYear))),
       amount => {
         pensionSessionService.getPensionsSessionDataResult(taxYear, request.user) {
           data =>
-            val pensionsCYAModel: PensionsCYAModel = data.map(_.pensions).getOrElse(
-              PensionsCYAModel(PaymentsIntoPensionViewModel(), PensionAnnualAllowancesViewModel()))
+            val pensionsCYAModel: PensionsCYAModel = data.map(_.pensions).getOrElse(PensionsCYAModel.emptyModels)
             val viewModel: PaymentsIntoPensionViewModel = pensionsCYAModel.paymentsIntoPension
             val updatedCyaModel: PensionsCYAModel = {
-              pensionsCYAModel.copy(paymentsIntoPension = viewModel.copy(totalRASPaymentsAndTaxRelief = Some(amount)))
+              pensionsCYAModel.copy(paymentsIntoPension = viewModel.copy(totalRetirementAnnuityContractPayments = Some(amount)))
             }
             pensionSessionService.createOrUpdateSessionData(request.user,
               updatedCyaModel, taxYear, data.exists(_.isPriorSubmission))(errorHandler.internalServerError()) {
-              Redirect(ReliefAtSourceOneOffPaymentsController.show(taxYear))
+              Redirect(WorkplacePensionController.show(taxYear))
             }
         }
-
       }
     )
   }
