+++ app/controllers/pensions/paymentsIntoPension/WorkplaceAmountController.scala
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package controllers.pensions
+package controllers.pensions.paymentsIntoPension
 
 import config.{AppConfig, ErrorHandler}
-import controllers.pensions.routes._
+import controllers.pensions.paymentsIntoPension.routes._
 import controllers.predicates.AuthorisedAction
-import forms.{AmountForm, FormUtils}
+import forms.AmountForm
 import models.mongo.PensionsCYAModel
 import models.pension.charges.PensionAnnualAllowancesViewModel
 import models.pension.reliefs.PaymentsIntoPensionViewModel
@@ -28,40 +28,39 @@ import play.api.i18n.I18nSupport
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import services.PensionSessionService
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendController
-import utils.{Clock, SessionHelper}
-import views.html.pensions.RetirementAnnuityAmountView
+import utils.Clock
+import views.html.pensions.WorkplaceAmountView
 
-import javax.inject.{Inject, Singleton}
+import javax.inject.Inject
 import scala.concurrent.Future
 
+class WorkplaceAmountController @Inject()(implicit val cc: MessagesControllerComponents,
+                                          authAction: AuthorisedAction,
+                                          workplaceAmountView: WorkplaceAmountView,
+                                          appConfig: AppConfig,
+                                          pensionSessionService: PensionSessionService,
+                                          errorHandler: ErrorHandler,
+                                          clock: Clock) extends FrontendController(cc) with I18nSupport {
 
-@Singleton
-class RetirementAnnuityAmountController @Inject()(implicit val mcc: MessagesControllerComponents,
-                                                  authAction: AuthorisedAction,
-                                                  retirementAnnuityAmountView: RetirementAnnuityAmountView,
-                                                  appConfig: AppConfig,
-                                                  pensionSessionService: PensionSessionService,
-                                                  errorHandler: ErrorHandler,
-                                                  clock: Clock) extends FrontendController(mcc) with I18nSupport with SessionHelper with FormUtils {
 
   val amountForm: Form[BigDecimal] = AmountForm.amountForm(
-    emptyFieldKey = "pensions.retirementAnnuityAmount.error.noEntry",
-    wrongFormatKey = "pensions.retirementAnnuityAmount.error.incorrectFormat",
-    exceedsMaxAmountKey = "pensions.retirementAnnuityAmount.error.overMaximum"
+    emptyFieldKey = "pensions.workplaceAmount.error.noEntry",
+    wrongFormatKey = "pensions.workplaceAmount.error.incorrectFormat",
+    exceedsMaxAmountKey = "pensions.workplaceAmount.error.maxAmount"
   )
-
 
   def show(taxYear: Int): Action[AnyContent] = authAction.async { implicit request =>
     pensionSessionService.getPensionsSessionDataResult(taxYear, request.user) {
       case Some(data) =>
-        if (data.pensions.paymentsIntoPension.retirementAnnuityContractPaymentsQuestion.contains(true)) {
-          data.pensions.paymentsIntoPension.totalRetirementAnnuityContractPayments match {
+        if (data.pensions.paymentsIntoPension.workplacePensionPaymentsQuestion.contains(true)) {
+          data.pensions.paymentsIntoPension.totalWorkplacePensionPayments match {
             case Some(amount) =>
-              Future.successful(Ok(retirementAnnuityAmountView(amountForm.fill(amount), taxYear)))
-            case None => Future.successful(Ok(retirementAnnuityAmountView(amountForm, taxYear)))
+              Future.successful(Ok(workplaceAmountView(amountForm.fill(amount), taxYear)))
+            case None =>
+              Future.successful(Ok(workplaceAmountView(amountForm, taxYear)))
           }
         } else {
-          Future.successful(Redirect(RetirementAnnuityController.show(taxYear)))
+          Future.successful(Redirect(WorkplacePensionController.show(taxYear)))
         }
       case _ =>
         Future.successful(Redirect(PaymentsIntoPensionsCYAController.show(taxYear)))
@@ -69,26 +68,24 @@ class RetirementAnnuityAmountController @Inject()(implicit val mcc: MessagesCont
 
   }
 
-
   def submit(taxYear: Int): Action[AnyContent] = authAction.async { implicit request =>
     amountForm.bindFromRequest.fold(
-      formWithErrors => Future.successful(BadRequest(retirementAnnuityAmountView(formWithErrors, taxYear))),
+      formWithErrors => Future.successful(BadRequest(workplaceAmountView(formWithErrors, taxYear))),
       amount => {
         pensionSessionService.getPensionsSessionDataResult(taxYear, request.user) {
           data =>
-            val pensionsCYAModel: PensionsCYAModel = data.map(_.pensions).getOrElse(PensionsCYAModel(
-              PaymentsIntoPensionViewModel(), PensionAnnualAllowancesViewModel()))
+            val pensionsCYAModel: PensionsCYAModel = data.map(_.pensions).getOrElse(PensionsCYAModel.emptyModels)
             val viewModel: PaymentsIntoPensionViewModel = pensionsCYAModel.paymentsIntoPension
             val updatedCyaModel: PensionsCYAModel = {
-              pensionsCYAModel.copy(paymentsIntoPension = viewModel.copy(totalRetirementAnnuityContractPayments = Some(amount)))
+              pensionsCYAModel.copy(paymentsIntoPension = viewModel.copy(totalWorkplacePensionPayments = Some(amount)))
             }
             pensionSessionService.createOrUpdateSessionData(request.user,
               updatedCyaModel, taxYear, data.exists(_.isPriorSubmission))(errorHandler.internalServerError()) {
-              Redirect(WorkplacePensionController.show(taxYear))
+              Redirect(PaymentsIntoPensionsCYAController.show(taxYear))
             }
         }
+
       }
     )
   }
-
 }
