+++ app/views/pensions/PaymentsIntoPensionsCYAView.scala.html
@@ -19,7 +19,6 @@
 @import views.html.templates.helpers.Heading
 @import views.html.templates.helpers.Button
 @import models.pension.reliefs.PaymentsIntoPensionViewModel
-@import models.pension.AllPensionsData
 @import utils.ViewUtils.bigDecimalCurrency
 @import utils.ViewUtils.summaryListRow
 
@@ -56,15 +55,15 @@
 }
 
 @summaryRows = @{
-    (Nil :+ rowWithContent("paymentsIntoPensions.cya.reliefAtSource", cya.rasPensionPaymentQuestion.getOrElse(false), changeLink = controllers.pensions.routes.ReliefAtSourcePensionsController.show(taxYear))) :+
-            amountRow("paymentsIntoPensions.cya.reliefAtSourceTotal", cya.totalRASPaymentsAndTaxRelief, cya.rasPensionPaymentQuestion.contains(true), changeLink = controllers.pensions.routes.ReliefAtSourcePaymentsAndTaxReliefAmountController.show(taxYear)) :+
-            rowWithContent("paymentsIntoPensions.cya.oneOff", cya.oneOffRasPaymentPlusTaxReliefQuestion.getOrElse(false), cya.rasPensionPaymentQuestion.contains(true), changeLink = controllers.pensions.routes.ReliefAtSourceOneOffPaymentsController.show(taxYear)) :+
-            amountRow("paymentsIntoPensions.cya.oneOffTotal", cya.totalOneOffRasPaymentPlusTaxRelief, cya.oneOffRasPaymentPlusTaxReliefQuestion.contains(true), changeLink = controllers.pensions.routes.OneOffRASPaymentsAmountController.show(taxYear)) :+
-            rowWithContent("paymentsIntoPensions.cya.pensionsTaxReliefNotClaimed", cya.pensionTaxReliefNotClaimedQuestion.getOrElse(false), changeLink = controllers.pensions.routes.PensionsTaxReliefNotClaimedController.show(taxYear)) :+
-            rowWithContent("paymentsIntoPensions.cya.retirementAnnuity", cya.retirementAnnuityContractPaymentsQuestion.getOrElse(false), cya.pensionTaxReliefNotClaimedQuestion.getOrElse(false), changeLink = controllers.pensions.routes.RetirementAnnuityController.show(taxYear)) :+
-            amountRow("paymentsIntoPensions.cya.retirementAnnuityTotal", cya.totalRetirementAnnuityContractPayments, cya.retirementAnnuityContractPaymentsQuestion.getOrElse(false), changeLink = controllers.pensions.routes.RetirementAnnuityAmountController.show(taxYear)) :+
-            rowWithContent("paymentsIntoPensions.cya.workplacePayments", cya.workplacePensionPaymentsQuestion.getOrElse(false), cya.pensionTaxReliefNotClaimedQuestion.getOrElse(false), changeLink = controllers.pensions.routes.WorkplacePensionController.show(taxYear)) :+
-            amountRow("paymentsIntoPensions.cya.workplacePaymentsTotal", cya.totalWorkplacePensionPayments, cya.workplacePensionPaymentsQuestion.getOrElse(false), changeLink = controllers.pensions.routes.WorkplaceAmountController.show(taxYear))
+    (Nil :+ rowWithContent("paymentsIntoPensions.cya.reliefAtSource", cya.rasPensionPaymentQuestion.getOrElse(false), changeLink = controllers.pensions.paymentsIntoPension.routes.ReliefAtSourcePensionsController.show(taxYear))) :+
+            amountRow("paymentsIntoPensions.cya.reliefAtSourceTotal", cya.totalRASPaymentsAndTaxRelief, cya.rasPensionPaymentQuestion.contains(true), changeLink = controllers.pensions.paymentsIntoPension.routes.ReliefAtSourcePaymentsAndTaxReliefAmountController.show(taxYear)) :+
+            rowWithContent("paymentsIntoPensions.cya.oneOff", cya.oneOffRasPaymentPlusTaxReliefQuestion.getOrElse(false), cya.rasPensionPaymentQuestion.contains(true), changeLink = controllers.pensions.paymentsIntoPension.routes.ReliefAtSourceOneOffPaymentsController.show(taxYear)) :+
+            amountRow("paymentsIntoPensions.cya.oneOffTotal", cya.totalOneOffRasPaymentPlusTaxRelief, cya.oneOffRasPaymentPlusTaxReliefQuestion.contains(true), changeLink = controllers.pensions.paymentsIntoPension.routes.OneOffRASPaymentsAmountController.show(taxYear)) :+
+            rowWithContent("paymentsIntoPensions.cya.pensionsTaxReliefNotClaimed", cya.pensionTaxReliefNotClaimedQuestion.getOrElse(false), changeLink = controllers.pensions.paymentsIntoPension.routes.PensionsTaxReliefNotClaimedController.show(taxYear)) :+
+            rowWithContent("paymentsIntoPensions.cya.retirementAnnuity", cya.retirementAnnuityContractPaymentsQuestion.getOrElse(false), cya.pensionTaxReliefNotClaimedQuestion.getOrElse(false), changeLink = controllers.pensions.paymentsIntoPension.routes.RetirementAnnuityController.show(taxYear)) :+
+            amountRow("paymentsIntoPensions.cya.retirementAnnuityTotal", cya.totalRetirementAnnuityContractPayments, cya.retirementAnnuityContractPaymentsQuestion.getOrElse(false), changeLink = controllers.pensions.paymentsIntoPension.routes.RetirementAnnuityAmountController.show(taxYear)) :+
+            rowWithContent("paymentsIntoPensions.cya.workplacePayments", cya.workplacePensionPaymentsQuestion.getOrElse(false), cya.pensionTaxReliefNotClaimedQuestion.getOrElse(false), changeLink = controllers.pensions.paymentsIntoPension.routes.WorkplacePensionController.show(taxYear)) :+
+            amountRow("paymentsIntoPensions.cya.workplacePaymentsTotal", cya.totalWorkplacePensionPayments, cya.workplacePensionPaymentsQuestion.getOrElse(false), changeLink = controllers.pensions.paymentsIntoPension.routes.WorkplaceAmountController.show(taxYear))
 }
 
 @layout(pageTitle = headingForUse, taxYear = Some(taxYear)) {
@@ -73,7 +72,7 @@
 
     @summaryList(summaryRows.flatten)
 
-    @csrfForm(controllers.pensions.routes.PaymentsIntoPensionsCYAController.submit(taxYear)) {
+    @csrfForm(controllers.pensions.paymentsIntoPension.routes.PaymentsIntoPensionsCYAController.submit(taxYear)) {
         @button("common.saveAndContinue")
     }
 }
