+++ app/views/pensions/annual/AboveReducedAnnualAllowanceAmountView.scala.html
@@ -15,11 +15,11 @@
  *@
 
 @import views.html.templates.Layout
+@import views.html.templates.helpers.InputText
 @import views.html.templates.helpers.Heading
-@import views.html.templates.helpers.InputRadio
 @import views.html.templates.helpers.Button
 @import views.html.templates.helpers.ErrorSummary
-@import views.html.templates.helpers.InputText
+@import utils.ViewUtils._
 
 @this(
         layout: Layout,
@@ -27,55 +27,39 @@
         button: Button,
         formWithCsrf: FormWithCSRF,
         inputText: InputText,
-        inputRadio: InputRadio,
-        govukInsetText: GovukInsetText,
         errorSummary: ErrorSummary
 )
 
-
 @(form: Form[BigDecimal], taxYear: Int)(implicit request: AuthorisationRequest[_], messages: Messages, appConfig: AppConfig)
-
-@titleForUse = @{messages("pensions.reliefAtSourceTotalPaymentsAndTaxReliefAmount.title")}
+@titleText = @{
+    messages(s"pensions.reducedAnnualAllowanceAmount.title.${if(request.user.isAgent) "agent" else "individual"}")
+}
 
 @hasFormErrors = @{form.hasErrors}
 
 @contentHtml = {
-
-    <p class="govuk-body">@{messages(s"pensions.reliefAtSourceTotalPaymentsAndTaxReliefAmount.whereToFind.${if(request.user.isAgent) "agent" else "individual"}")}</p>
-    <p class="govuk-body">@{messages(s"pensions.reliefAtSourceTotalPaymentsAndTaxReliefAmount.howToWorkOut.${if(request.user.isAgent) "agent" else "individual"}")}</p>
-
-    <div class="govuk-inset-text">
-        <span class="govuk-form-group ">
-            <strong>@messages("paymentsIntoPensions.reliefAtSource.calculationHeading")</strong>
-        </span>
-        <span>@messages("pensions.reliefAtSourceTotalPaymentsAndTaxReliefAmount.exampleCalculation")</span>
-    </div>
-
+    <p class="govuk-body">@{messages(s"pensions.reducedAnnualAllowanceAmount.paragraph")}</p>
 }
 
-
-@layout(pageTitle = titleForUse, taxYear = Some(taxYear), hasErrors = hasFormErrors, isAgent = request.user.isAgent) {
+@layout(pageTitle = titleText, taxYear = Some(taxYear), hasErrors = hasFormErrors, isAgent = request.user.isAgent) {
 
     @errorSummary(form.errors.distinct)
 
-    @formWithCsrf(action = controllers.pensions.paymentsIntoPension.routes.ReliefAtSourcePaymentsAndTaxReliefAmountController.submit(taxYear)) {
-
+    @formWithCsrf(action = controllers.pensions.annualAllowance.routes.AboveReducedAnnualAllowanceAmountController.submit(taxYear)) {
         @inputText(
             form,
             id = "amount",
             name = "amount",
-            heading = titleForUse,
+            heading = titleText,
             headingClasses = "govuk-!-margin-bottom-2",
-            subheading = Some(messages("common.paymentsIntoPensions.caption", (taxYear - 1).toString, taxYear.toString)),
+            subheading = Some(messages("common.pensionAnnualAllowance.caption", (taxYear - 1).toString, taxYear.toString)),
             content = Some(contentHtml),
             hint = Some(Html(messages("common.currency.hint"))),
             isPageHeading = false,
             currency = true
         )
-
         @button()
     }
-
 }
 
 @{
