+++ it/controllers/pensions/annualAllowance/AboveReducedAnnualAllowanceAmountControllerISpec.scala
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package controllers.pensions
+package controllers.pensions.annualAllowance
 
-import builders.PaymentsIntoPensionVewModelBuilder.aPaymentsIntoPensionViewModel
+import builders.PensionAnnualAllowanceViewModelBuilder.aPensionAnnualAllowanceViewModel
 import builders.PensionsCYAModelBuilder._
 import builders.PensionsUserDataBuilder
 import builders.UserBuilder._
@@ -28,11 +28,12 @@ import org.scalatest.BeforeAndAfterEach
 import play.api.http.HeaderNames
 import play.api.http.Status.{BAD_REQUEST, OK, SEE_OTHER}
 import play.api.libs.ws.WSResponse
-import utils.PageUrls.{checkPaymentsIntoPensionCyaUrl, fullUrl, retirementAnnuityAmountUrl, retirementAnnuityUrl, workplacePensionUrl}
+import utils.PageUrls.PensionAnnualAllowancePages.aboveReducedAnnualAllowanceAmountUrl
+import utils.PageUrls.{fullUrl, pensionSummaryUrl}
 import utils.{IntegrationTest, PensionsDatabaseHelper, ViewHelpers}
 
 
-class RetirementAnnuityAmountControllerISpec extends IntegrationTest with ViewHelpers with BeforeAndAfterEach with PensionsDatabaseHelper {
+class AboveReducedAnnualAllowanceAmountControllerISpec extends IntegrationTest with ViewHelpers with BeforeAndAfterEach with PensionsDatabaseHelper {
   private val taxYearEOY: Int = taxYear - 1
   private val poundPrefixText = "£"
   private val amountInputName = "amount"
@@ -59,60 +60,56 @@ class RetirementAnnuityAmountControllerISpec extends IntegrationTest with ViewHe
     val maxAmountErrorText: String
     val hintText: String
     val buttonText: String
+    val expectedParagraph: String
   }
 
   trait SpecificExpectedResults {
-    val onlyIncludePayment: String
     val expectedTitle: String
     val expectedHeading: String
     val expectedErrorTitle: String
   }
 
   object CommonExpectedEN extends CommonExpectedResults {
-    val expectedCaption: Int => String = (taxYear: Int) => s"Payments into pensions for 6 April ${taxYear - 1} to 5 April $taxYear"
+    val expectedCaption: Int => String = (taxYear: Int) => s"Pension annual allowance for 6 April ${taxYear - 1} to 5 April $taxYear"
     val hintText = "For example, £193.52"
-    val emptyErrorText = "Enter the amount paid into retirement annuity contracts"
-    val invalidFormatErrorText = "Enter the amount paid into retirement annuity contracts in the correct format"
-    val maxAmountErrorText = "The amount paid into retirement annuity contracts must be less than £100,000,000,000"
+    val emptyErrorText = "Enter the amount above your reduced annual allowance"
+    val invalidFormatErrorText = "Enter the amount above your reduced annual allowance in the correct format"
+    val maxAmountErrorText = "The amount above your reduced annual allowance must be less than £100,000,000,000"
     val buttonText = "Continue"
+    val expectedParagraph = "This is the amount on which tax is due."
   }
 
   object CommonExpectedCY extends CommonExpectedResults {
-    val expectedCaption: Int => String = (taxYear: Int) => s"Payments into pensions for 6 April ${taxYear - 1} to 5 April $taxYear"
+    val expectedCaption: Int => String = (taxYear: Int) => s"Pension annual allowance for 6 April ${taxYear - 1} to 5 April $taxYear"
     val hintText = "For example, £193.52"
-    val emptyErrorText = "Enter the amount paid into retirement annuity contracts"
-    val invalidFormatErrorText = "Enter the amount paid into retirement annuity contracts in the correct format"
-    val maxAmountErrorText = "The amount paid into retirement annuity contracts must be less than £100,000,000,000"
+    val emptyErrorText = "Enter the amount above your reduced annual allowance"
+    val invalidFormatErrorText = "Enter the amount above your reduced annual allowance in the correct format"
+    val maxAmountErrorText = "The amount above your reduced annual allowance must be less than £100,000,000,000"
     val buttonText = "Continue"
+    val expectedParagraph = "This is the amount on which tax is due."
   }
 
   object ExpectedIndividualEN extends SpecificExpectedResults {
-    val onlyIncludePayment = "Only include payments your pension provider will not claim tax relief for. You can find this out from your pension provider."
-    val expectedTitle = "How much did you pay into your retirement annuity contracts?"
-    val expectedHeading = "How much did you pay into your retirement annuity contracts?"
+    val expectedTitle = "How much above your annual allowance are you?"
+    val expectedHeading = "How much above your annual allowance are you?"
     val expectedErrorTitle = s"Error: $expectedTitle"
   }
 
   object ExpectedIndividualCY extends SpecificExpectedResults {
-    val onlyIncludePayment = "Only include payments your pension provider will not claim tax relief for. You can find this out from your pension provider."
-    val expectedTitle = "How much did you pay into your retirement annuity contracts?"
-    val expectedHeading = "How much did you pay into your retirement annuity contracts?"
+    val expectedTitle = "How much above your annual allowance are you?"
+    val expectedHeading = "How much above your annual allowance are you?"
     val expectedErrorTitle = s"Error: $expectedTitle"
   }
 
   object ExpectedAgentEN extends SpecificExpectedResults {
-    val onlyIncludePayment =
-      "Only include payments your client’s pension provider will not claim tax relief for. You can find this out from your client’s pension provider."
-    val expectedTitle = "How much did your client pay into their retirement annuity contracts?"
-    val expectedHeading = "How much did your client pay into their retirement annuity contracts?"
+    val expectedTitle = "How much above your client’s reduced annual allowance are they?"
+    val expectedHeading = "How much above your client’s reduced annual allowance are they?"
     val expectedErrorTitle = s"Error: $expectedTitle"
   }
 
   object ExpectedAgentCY extends SpecificExpectedResults {
-    val onlyIncludePayment =
-      "Only include payments your client’s pension provider will not claim tax relief for. You can find this out from your client’s pension provider."
-    val expectedTitle = "How much did your client pay into their retirement annuity contracts?"
-    val expectedHeading = "How much did your client pay into their retirement annuity contracts?"
+    val expectedTitle = "How much above your client’s reduced annual allowance are they?"
+    val expectedHeading = "How much above your client’s reduced annual allowance are they?"
     val expectedErrorTitle = s"Error: $expectedTitle"
   }
 
@@ -128,14 +125,14 @@ class RetirementAnnuityAmountControllerISpec extends IntegrationTest with ViewHe
       import user.commonExpectedResults._
 
       s"language is ${welshTest(user.isWelsh)} and request is from an ${agentTest(user.isAgent)}" should {
-        "render How much did you pay into your retirement annuity contracts page with no value when no cya data" which {
+        "render How much above your annual allowance are you page with no value when no cya data" which {
           lazy val result: WSResponse = {
             dropPensionsDB()
             authoriseAgentOrIndividual(user.isAgent)
-            val pensionsViewModel = aPaymentsIntoPensionViewModel.copy(
-              retirementAnnuityContractPaymentsQuestion = Some(true), totalRetirementAnnuityContractPayments = None)
-            insertCyaData(pensionsUsersData(isPrior = false, aPensionsCYAModel.copy(paymentsIntoPension = pensionsViewModel)), aUserRequest)
-            urlGet(fullUrl(retirementAnnuityAmountUrl(taxYearEOY)), user.isWelsh, follow = false,
+            val pensionsViewModel = aPensionAnnualAllowanceViewModel.copy(
+              aboveAnnualAllowanceQuestion = Some(true), aboveAnnualAllowance = None)
+            insertCyaData(pensionsUsersData(isPrior = false, aPensionsCYAModel.copy(pensionsAnnualAllowances = pensionsViewModel)), aUserRequest)
+            urlGet(fullUrl(aboveReducedAnnualAllowanceAmountUrl(taxYearEOY)), user.isWelsh, follow = false,
               headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
           }
 
@@ -147,25 +144,25 @@ class RetirementAnnuityAmountControllerISpec extends IntegrationTest with ViewHe
           titleCheck(user.specificExpectedResults.get.expectedTitle)
           h1Check(user.specificExpectedResults.get.expectedHeading)
           captionCheck(expectedCaption(taxYearEOY), captionSelector)
-          textOnPageCheck(user.specificExpectedResults.get.onlyIncludePayment, paragraphSelector)
+          textOnPageCheck(expectedParagraph, paragraphSelector)
           textOnPageCheck(hintText, hintTextSelector)
           textOnPageCheck(poundPrefixText, poundPrefixSelector)
           inputFieldValueCheck(amountInputName, inputSelector, "")
           buttonCheck(buttonText, continueButtonSelector)
-          formPostLinkCheck(retirementAnnuityAmountUrl(taxYearEOY), formSelector)
+          formPostLinkCheck(aboveReducedAnnualAllowanceAmountUrl(taxYearEOY), formSelector)
           welshToggleCheck(user.isWelsh)
         }
 
-        "render How much did you pay into your retirement annuity contracts page prefilled when cya data" which {
+        "render How much above your annual allowance are you page prefilled when cya data" which {
 
           val existingAmount: String = "999.88"
           lazy val result: WSResponse = {
             dropPensionsDB()
             authoriseAgentOrIndividual(user.isAgent)
-            val pensionsViewModel = aPaymentsIntoPensionViewModel.copy(
-              retirementAnnuityContractPaymentsQuestion = Some(true), totalRetirementAnnuityContractPayments = Some(BigDecimal(existingAmount)))
-            insertCyaData(pensionsUsersData(isPrior = false, aPensionsCYAModel.copy(paymentsIntoPension = pensionsViewModel)), aUserRequest)
-            urlGet(fullUrl(retirementAnnuityAmountUrl(taxYearEOY)), user.isWelsh, follow = false,
+            val pensionsViewModel = aPensionAnnualAllowanceViewModel.copy(
+              aboveAnnualAllowanceQuestion = Some(true), aboveAnnualAllowance = Some(BigDecimal(existingAmount)))
+            insertCyaData(pensionsUsersData(isPrior = false, aPensionsCYAModel.copy(pensionsAnnualAllowances = pensionsViewModel)), aUserRequest)
+            urlGet(fullUrl(aboveReducedAnnualAllowanceAmountUrl(taxYearEOY)), user.isWelsh, follow = false,
               headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
           }
 
@@ -177,25 +174,26 @@ class RetirementAnnuityAmountControllerISpec extends IntegrationTest with ViewHe
           titleCheck(user.specificExpectedResults.get.expectedTitle)
           h1Check(user.specificExpectedResults.get.expectedHeading)
           captionCheck(expectedCaption(taxYearEOY), captionSelector)
-          textOnPageCheck(user.specificExpectedResults.get.onlyIncludePayment, paragraphSelector)
+          textOnPageCheck(expectedParagraph, paragraphSelector)
           textOnPageCheck(hintText, hintTextSelector)
           textOnPageCheck(poundPrefixText, poundPrefixSelector)
           inputFieldValueCheck(amountInputName, inputSelector, existingAmount)
           buttonCheck(buttonText, continueButtonSelector)
-          formPostLinkCheck(retirementAnnuityAmountUrl(taxYearEOY), formSelector)
+          formPostLinkCheck(aboveReducedAnnualAllowanceAmountUrl(taxYearEOY), formSelector)
           welshToggleCheck(user.isWelsh)
         }
       }
     }
 
-    "redirect to the retirementAnnuityContractQuestion page if the question has not been answered" which {
+    //TODO redirect to question page
+    "redirect to the PensionSummary page if the question has not been answered" which {
       lazy val result: WSResponse = {
         dropPensionsDB()
         authoriseAgentOrIndividual(isAgent = false)
-        val pensionsViewModel = aPaymentsIntoPensionViewModel.copy(
-          retirementAnnuityContractPaymentsQuestion = None, totalRetirementAnnuityContractPayments = None)
-        insertCyaData(pensionsUsersData(isPrior = false, aPensionsCYAModel.copy(paymentsIntoPension = pensionsViewModel)), aUserRequest)
-        urlGet(fullUrl(retirementAnnuityAmountUrl(taxYearEOY)), follow = false,
+        val pensionsViewModel = aPensionAnnualAllowanceViewModel.copy(
+          aboveAnnualAllowanceQuestion = None, aboveAnnualAllowance = None)
+        insertCyaData(pensionsUsersData(isPrior = false, aPensionsCYAModel.copy(pensionsAnnualAllowances = pensionsViewModel)), aUserRequest)
+        urlGet(fullUrl(aboveReducedAnnualAllowanceAmountUrl(taxYearEOY)), follow = false,
           headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
       }
 
@@ -203,18 +201,19 @@ class RetirementAnnuityAmountControllerISpec extends IntegrationTest with ViewHe
 
       "has an SEE_OTHER status" in {
         result.status shouldBe SEE_OTHER
-        result.header("location").contains(retirementAnnuityUrl(taxYearEOY)) shouldBe true
+        result.header("location").contains(pensionSummaryUrl(taxYearEOY)) shouldBe true
       }
     }
 
-    "redirect to the retirementAnnuityContractQuestion page if the retirementAnnuityContract question has been answered as false" which {
+    //TODO redirect to question page
+    "redirect to the PensionSummary page if the retirementAnnuityContract question has been answered as false" which {
       lazy val result: WSResponse = {
         dropPensionsDB()
         authoriseAgentOrIndividual(isAgent = false)
-        val pensionsViewModel = aPaymentsIntoPensionViewModel.copy(
-          retirementAnnuityContractPaymentsQuestion = Some(false), totalRetirementAnnuityContractPayments = None)
-        insertCyaData(pensionsUsersData(isPrior = false, aPensionsCYAModel.copy(paymentsIntoPension = pensionsViewModel)), aUserRequest)
-        urlGet(fullUrl(retirementAnnuityAmountUrl(taxYearEOY)), follow = false,
+        val pensionsViewModel = aPensionAnnualAllowanceViewModel.copy(
+          aboveAnnualAllowanceQuestion = Some(false), aboveAnnualAllowance = None)
+        insertCyaData(pensionsUsersData(isPrior = false, aPensionsCYAModel.copy(pensionsAnnualAllowances = pensionsViewModel)), aUserRequest)
+        urlGet(fullUrl(aboveReducedAnnualAllowanceAmountUrl(taxYearEOY)), follow = false,
           headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
       }
 
@@ -222,20 +221,21 @@ class RetirementAnnuityAmountControllerISpec extends IntegrationTest with ViewHe
 
       "has an SEE_OTHER status" in {
         result.status shouldBe SEE_OTHER
-        result.header("location").contains(retirementAnnuityUrl(taxYearEOY)) shouldBe true
+        result.header("location").contains(pensionSummaryUrl(taxYearEOY)) shouldBe true
       }
     }
 
-    "redirect to the CYA page if there is no session data" which {
+    //TODO redirect to annual CYA Page
+    "redirect to the Pension Summary page if there is no session data" which {
       lazy val result: WSResponse = {
         dropPensionsDB()
         authoriseAgentOrIndividual(isAgent = false)
-        urlGet(fullUrl(retirementAnnuityAmountUrl(taxYearEOY)), follow = false,
+        urlGet(fullUrl(aboveReducedAnnualAllowanceAmountUrl(taxYearEOY)), follow = false,
           headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
       }
       "has an SEE_OTHER status" in {
         result.status shouldBe SEE_OTHER
-        result.header("location").contains(checkPaymentsIntoPensionCyaUrl(taxYearEOY)) shouldBe true
+        result.header("location").contains(pensionSummaryUrl(taxYearEOY)) shouldBe true
       }
     }
   }
@@ -251,11 +251,11 @@ class RetirementAnnuityAmountControllerISpec extends IntegrationTest with ViewHe
           val emptyForm: Map[String, String] = Map(AmountForm.amount -> amountEmpty)
           lazy val result: WSResponse = {
             dropPensionsDB()
-            val pensionsViewModel = aPaymentsIntoPensionViewModel.copy(
-              retirementAnnuityContractPaymentsQuestion = Some(true), totalRetirementAnnuityContractPayments = None)
-            insertCyaData(pensionsUsersData(isPrior = false, aPensionsCYAModel.copy(paymentsIntoPension = pensionsViewModel)), aUserRequest)
+            val pensionsViewModel = aPensionAnnualAllowanceViewModel.copy(
+              aboveAnnualAllowanceQuestion = Some(true), aboveAnnualAllowance = None)
+            insertCyaData(pensionsUsersData(isPrior = false, aPensionsCYAModel.copy(pensionsAnnualAllowances = pensionsViewModel)), aUserRequest)
             authoriseAgentOrIndividual(user.isAgent)
-            urlPost(fullUrl(retirementAnnuityAmountUrl(taxYearEOY)), body = emptyForm, welsh = user.isWelsh,
+            urlPost(fullUrl(aboveReducedAnnualAllowanceAmountUrl(taxYearEOY)), body = emptyForm, welsh = user.isWelsh,
               follow = false, headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
           }
           "has the correct status" in {
@@ -267,12 +267,12 @@ class RetirementAnnuityAmountControllerISpec extends IntegrationTest with ViewHe
           titleCheck(user.specificExpectedResults.get.expectedErrorTitle)
           h1Check(user.specificExpectedResults.get.expectedHeading)
           captionCheck(expectedCaption(taxYearEOY), captionSelector)
-          textOnPageCheck(user.specificExpectedResults.get.onlyIncludePayment, paragraphSelector)
+          textOnPageCheck(expectedParagraph, paragraphSelector)
           textOnPageCheck(hintText, hintTextSelector)
           textOnPageCheck(poundPrefixText, poundPrefixSelector)
           inputFieldValueCheck(amountInputName, inputSelector, amountEmpty)
           buttonCheck(buttonText, continueButtonSelector)
-          formPostLinkCheck(retirementAnnuityAmountUrl(taxYearEOY), formSelector)
+          formPostLinkCheck(aboveReducedAnnualAllowanceAmountUrl(taxYearEOY), formSelector)
           errorSummaryCheck(emptyErrorText, expectedErrorHref)
           errorAboveElementCheck(emptyErrorText)
           welshToggleCheck(user.isWelsh)
@@ -285,11 +285,11 @@ class RetirementAnnuityAmountControllerISpec extends IntegrationTest with ViewHe
 
           lazy val result: WSResponse = {
             dropPensionsDB()
-            val pensionsViewModel = aPaymentsIntoPensionViewModel.copy(
-              retirementAnnuityContractPaymentsQuestion = Some(true), totalRetirementAnnuityContractPayments = None)
-            insertCyaData(pensionsUsersData(isPrior = false, aPensionsCYAModel.copy(paymentsIntoPension = pensionsViewModel)), aUserRequest)
+            val pensionsViewModel = aPensionAnnualAllowanceViewModel.copy(
+              aboveAnnualAllowanceQuestion = Some(true), aboveAnnualAllowance = None)
+            insertCyaData(pensionsUsersData(isPrior = false, aPensionsCYAModel.copy(pensionsAnnualAllowances = pensionsViewModel)), aUserRequest)
             authoriseAgentOrIndividual(user.isAgent)
-            urlPost(fullUrl(retirementAnnuityAmountUrl(taxYearEOY)), body = invalidFormatForm, welsh = user.isWelsh,
+            urlPost(fullUrl(aboveReducedAnnualAllowanceAmountUrl(taxYearEOY)), body = invalidFormatForm, welsh = user.isWelsh,
               follow = false, headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
           }
 
@@ -302,12 +302,12 @@ class RetirementAnnuityAmountControllerISpec extends IntegrationTest with ViewHe
           titleCheck(user.specificExpectedResults.get.expectedErrorTitle)
           h1Check(user.specificExpectedResults.get.expectedHeading)
           captionCheck(expectedCaption(taxYearEOY), captionSelector)
-          textOnPageCheck(user.specificExpectedResults.get.onlyIncludePayment, paragraphSelector)
+          textOnPageCheck(expectedParagraph, paragraphSelector)
           textOnPageCheck(hintText, hintTextSelector)
           textOnPageCheck(poundPrefixText, poundPrefixSelector)
           inputFieldValueCheck(amountInputName, inputSelector, amountInvalidFormat)
           buttonCheck(buttonText, continueButtonSelector)
-          formPostLinkCheck(retirementAnnuityAmountUrl(taxYearEOY), formSelector)
+          formPostLinkCheck(aboveReducedAnnualAllowanceAmountUrl(taxYearEOY), formSelector)
           errorSummaryCheck(invalidFormatErrorText, expectedErrorHref)
           errorAboveElementCheck(invalidFormatErrorText)
           welshToggleCheck(user.isWelsh)
@@ -320,11 +320,11 @@ class RetirementAnnuityAmountControllerISpec extends IntegrationTest with ViewHe
 
           lazy val result: WSResponse = {
             dropPensionsDB()
-            val pensionsViewModel = aPaymentsIntoPensionViewModel.copy(
-              retirementAnnuityContractPaymentsQuestion = Some(true), totalRetirementAnnuityContractPayments = None)
-            insertCyaData(pensionsUsersData(isPrior = false, aPensionsCYAModel.copy(paymentsIntoPension = pensionsViewModel)), aUserRequest)
+            val pensionsViewModel = aPensionAnnualAllowanceViewModel.copy(
+              aboveAnnualAllowanceQuestion = Some(true), aboveAnnualAllowance = None)
+            insertCyaData(pensionsUsersData(isPrior = false, aPensionsCYAModel.copy(pensionsAnnualAllowances = pensionsViewModel)), aUserRequest)
             authoriseAgentOrIndividual(user.isAgent)
-            urlPost(fullUrl(retirementAnnuityAmountUrl(taxYearEOY)),
+            urlPost(fullUrl(aboveReducedAnnualAllowanceAmountUrl(taxYearEOY)),
               body = overMaximumForm, welsh = user.isWelsh, follow = false, headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
           }
           "has the correct status" in {
@@ -336,12 +336,12 @@ class RetirementAnnuityAmountControllerISpec extends IntegrationTest with ViewHe
           titleCheck(user.specificExpectedResults.get.expectedErrorTitle)
           h1Check(user.specificExpectedResults.get.expectedHeading)
           captionCheck(expectedCaption(taxYearEOY), captionSelector)
-          textOnPageCheck(user.specificExpectedResults.get.onlyIncludePayment, paragraphSelector)
+          textOnPageCheck(expectedParagraph, paragraphSelector)
           textOnPageCheck(hintText, hintTextSelector)
           textOnPageCheck(poundPrefixText, poundPrefixSelector)
           inputFieldValueCheck(amountInputName, inputSelector, amountOverMaximum)
           buttonCheck(buttonText, continueButtonSelector)
-          formPostLinkCheck(retirementAnnuityAmountUrl(taxYearEOY), formSelector)
+          formPostLinkCheck(aboveReducedAnnualAllowanceAmountUrl(taxYearEOY), formSelector)
           errorSummaryCheck(maxAmountErrorText, expectedErrorHref)
           errorAboveElementCheck(maxAmountErrorText)
           welshToggleCheck(user.isWelsh)
@@ -349,6 +349,7 @@ class RetirementAnnuityAmountControllerISpec extends IntegrationTest with ViewHe
       }
     }
 
+    //TODO: redirect to "Has your client’s pension provider paid the annual allowance tax?" page
     "redirect to the correct page when a valid amount is submitted and update the session amount" which {
 
       val validAmount = "1888.88"
@@ -356,23 +357,23 @@ class RetirementAnnuityAmountControllerISpec extends IntegrationTest with ViewHe
 
       lazy val result: WSResponse = {
         dropPensionsDB()
-        val pensionsViewModel = aPaymentsIntoPensionViewModel.copy(
-          retirementAnnuityContractPaymentsQuestion = Some(true), totalRetirementAnnuityContractPayments = None)
-        insertCyaData(pensionsUsersData(isPrior = false, aPensionsCYAModel.copy(paymentsIntoPension = pensionsViewModel)), aUserRequest)
+        val pensionsViewModel = aPensionAnnualAllowanceViewModel.copy(
+          aboveAnnualAllowanceQuestion = Some(true), aboveAnnualAllowance = None)
+        insertCyaData(pensionsUsersData(isPrior = false, aPensionsCYAModel.copy(pensionsAnnualAllowances = pensionsViewModel)), aUserRequest)
         authoriseAgentOrIndividual(isAgent = false)
-        urlPost(fullUrl(retirementAnnuityAmountUrl(taxYearEOY)),
+        urlPost(fullUrl(aboveReducedAnnualAllowanceAmountUrl(taxYearEOY)),
           body = validForm, follow = false,
           headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
       }
 
       "has a SEE_OTHER(303) status" in {
         result.status shouldBe SEE_OTHER
-        result.header("location") shouldBe Some(workplacePensionUrl(taxYearEOY))
+        result.header("location") shouldBe Some(pensionSummaryUrl(taxYearEOY))
       }
 
       "updates retirement annuity contract payments question to Some(true)" in {
         lazy val cyaModel = findCyaData(taxYearEOY, aUserRequest).get
-        cyaModel.pensions.paymentsIntoPension.totalRetirementAnnuityContractPayments shouldBe Some(BigDecimal(validAmount))
+        cyaModel.pensions.pensionsAnnualAllowances.aboveAnnualAllowance shouldBe Some(BigDecimal(validAmount))
       }
     }
   }
