+++ it/controllers/pensions/paymentsIntoPensions/WorkplaceAmountControllerISpec.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package controllers.pensions
+package controllers.pensions.paymentsIntoPensions
 
 import builders.PaymentsIntoPensionVewModelBuilder.aPaymentsIntoPensionViewModel
 import builders.PensionsCYAModelBuilder._
@@ -28,10 +28,12 @@ import org.scalatest.BeforeAndAfterEach
 import play.api.http.HeaderNames
 import play.api.http.Status.{BAD_REQUEST, OK, SEE_OTHER}
 import play.api.libs.ws.WSResponse
-import utils.PageUrls._
+import utils.PageUrls.PaymentIntoPensions.{checkPaymentsIntoPensionCyaUrl, workplacePensionAmount, workplacePensionUrl}
+import utils.PageUrls.fullUrl
 import utils.{IntegrationTest, PensionsDatabaseHelper, ViewHelpers}
 
-class ReliefAtSourcePaymentsAndTaxReliefAmountControllerISpec extends IntegrationTest with ViewHelpers with BeforeAndAfterEach with PensionsDatabaseHelper {
+// scalastyle:off magic.number
+class WorkplaceAmountControllerISpec extends IntegrationTest with ViewHelpers with BeforeAndAfterEach with PensionsDatabaseHelper {
 
   private val taxYearEOY: Int = taxYear - 1
   private val poundPrefixText = "£"
@@ -52,83 +54,89 @@ class ReliefAtSourcePaymentsAndTaxReliefAmountControllerISpec extends Integratio
     val poundPrefixSelector = ".govuk-input__prefix"
     val inputSelector = "#amount"
     val expectedErrorHref = "#amount"
+
+    def bulletListSelector(index: Int): String = s"#main-content > div > div > form > div > label > ul > li:nth-child($index)"
+
     def insetSpanText(index: Int): String = s"#main-content > div > div > form > div > label > div > span:nth-child($index)"
+
     def paragraphSelector(index: Int): String = s"#main-content > div > div > form > div > label > p:nth-child($index)"
   }
 
   trait CommonExpectedResults {
     val expectedCaption: Int => String
-    val expectedHeading: String
-    val expectedTitle: String
-    val expectedErrorTitle: String
-    val expectedCalculationHeading: String
-    val expectedExampleCalculation: String
     val emptyErrorText: String
     val invalidFormatErrorText: String
     val maxAmountErrorText: String
+    val expectedParagraph: String
     val hintText: String
     val buttonText: String
+
   }
 
   trait SpecificExpectedResults {
-    val expectedWhereToFind: String
-    val expectedHowToWorkOut: String
+    val expectedHeading: String
+    val expectedTitle: String
+    val expectedErrorTitle: String
+    val expectedBullet1: String
+    val expectedBullet2: String
+    val expectedYouCanFindThisOut: String
   }
 
   object CommonExpectedEN extends CommonExpectedResults {
     val expectedCaption: Int => String = (taxYear: Int) => s"Payments into pensions for 6 April ${taxYear - 1} to 5 April $taxYear"
-    val expectedHeading = "Total payments into relief at source (RAS) pensions, plus basic rate tax relief"
-    val expectedTitle = "Total payments into relief at source (RAS) pensions, plus basic rate tax relief"
-    val expectedErrorTitle = s"Error: $expectedTitle"
-    val expectedCalculationHeading = "Example calculation"
-    val expectedExampleCalculation = "Emma paid £500 into her pension scheme. £500 divided by 80 and multiplied by 100 is £625. Her answer is £625."
     val hintText = "For example, £193.52"
-    val emptyErrorText = "Enter the total paid into RAS pensions, plus basic rate tax relief"
-    val invalidFormatErrorText = "Enter the total paid into RAS pensions, plus basic rate tax relief, in the correct format"
-    val maxAmountErrorText = "The total paid into RAS pensions, plus basic rate tax relief, must be less than £100,000,000,000"
+    val emptyErrorText = "Enter the amount paid into workplace pensions"
+    val invalidFormatErrorText = "Enter the amount paid into workplace pensions in the correct format"
+    val maxAmountErrorText = "The amount paid into workplace pensions must be less than £100,000,000,000"
     val buttonText = "Continue"
+    val expectedParagraph = "Only include payments:"
   }
 
   object CommonExpectedCY extends CommonExpectedResults {
     val expectedCaption: Int => String = (taxYear: Int) => s"Payments into pensions for 6 April ${taxYear - 1} to 5 April $taxYear"
-    val expectedHeading = "Total payments into relief at source (RAS) pensions, plus basic rate tax relief"
-    val expectedTitle = "Total payments into relief at source (RAS) pensions, plus basic rate tax relief"
-    val expectedErrorTitle = s"Error: $expectedTitle"
-    val expectedCalculationHeading = "Example calculation"
-    val expectedExampleCalculation = "Emma paid £500 into her pension scheme. £500 divided by 80 and multiplied by 100 is £625. Her answer is £625."
     val hintText = "For example, £193.52"
-    val emptyErrorText = "Enter the total paid into RAS pensions, plus basic rate tax relief"
-    val invalidFormatErrorText = "Enter the total paid into RAS pensions, plus basic rate tax relief, in the correct format"
-    val maxAmountErrorText = "The total paid into RAS pensions, plus basic rate tax relief, must be less than £100,000,000,000"
+    val emptyErrorText = "Enter the amount paid into workplace pensions"
+    val invalidFormatErrorText = "Enter the amount paid into workplace pensions in the correct format"
+    val maxAmountErrorText = "The amount paid into workplace pensions must be less than £100,000,000,000"
     val buttonText = "Continue"
+    val expectedParagraph = "Only include payments:"
   }
 
   object ExpectedIndividualEN extends SpecificExpectedResults {
-    val expectedWhereToFind =
-      "You can find the total amount you paid into RAS pensions, plus tax relief, on the pension certificate or receipt from your administrator."
-    val expectedHowToWorkOut =
-      "To work it out yourself, divide the amount you actually paid by 80 and multiply the result by 100."
+    val expectedHeading = "How much did you pay into your workplace pensions?"
+    val expectedTitle = "How much did you pay into your workplace pensions?"
+    val expectedErrorTitle = s"Error: $expectedTitle"
+    val expectedBullet1 = "made after your pay was taxed"
+    val expectedBullet2 = "your pension provider will not claim tax relief for"
+    val expectedYouCanFindThisOut = "You can find this out from your employer or your pension provider."
   }
 
   object ExpectedIndividualCY extends SpecificExpectedResults {
-    val expectedWhereToFind =
-      "You can find the total amount you paid into RAS pensions, plus tax relief, on the pension certificate or receipt from your administrator."
-    val expectedHowToWorkOut =
-      "To work it out yourself, divide the amount you actually paid by 80 and multiply the result by 100."
+    val expectedHeading = "How much did you pay into your workplace pensions?"
+    val expectedTitle = "How much did you pay into your workplace pensions?"
+    val expectedErrorTitle = s"Error: $expectedTitle"
+    val expectedBullet1 = "made after your pay was taxed"
+    val expectedBullet2 = "your pension provider will not claim tax relief for"
+    val expectedYouCanFindThisOut = "You can find this out from your employer or your pension provider."
+
   }
 
   object ExpectedAgentEN extends SpecificExpectedResults {
-    val expectedWhereToFind =
-      "You can find the total amount your client paid into RAS pensions, plus tax relief, on the pension certificate or receipt from your administrator."
-    val expectedHowToWorkOut =
-      "To work it out yourself, divide the amount your client actually paid by 80 and multiply the result by 100."
+    val expectedHeading = "How much did your client pay into their workplace pensions?"
+    val expectedTitle = "How much did your client pay into their workplace pensions?"
+    val expectedErrorTitle = s"Error: $expectedTitle"
+    val expectedBullet1 = "made after your client’s pay was taxed"
+    val expectedBullet2 = "your client’s pension provider will not claim tax relief for"
+    val expectedYouCanFindThisOut = "Your client can find this out from their employer or pension provider."
   }
 
   object ExpectedAgentCY extends SpecificExpectedResults {
-    val expectedWhereToFind =
-      "You can find the total amount your client paid into RAS pensions, plus tax relief, on the pension certificate or receipt from your administrator."
-    val expectedHowToWorkOut =
-      "To work it out yourself, divide the amount your client actually paid by 80 and multiply the result by 100."
+    val expectedHeading = "How much did your client pay into their workplace pensions?"
+    val expectedTitle = "How much did your client pay into their workplace pensions?"
+    val expectedErrorTitle = s"Error: $expectedTitle"
+    val expectedBullet1 = "made after your client’s pay was taxed"
+    val expectedBullet2 = "your client’s pension provider will not claim tax relief for"
+    val expectedYouCanFindThisOut = "Your client can find this out from their employer or pension provider."
   }
 
   val userScenarios: Seq[UserScenario[CommonExpectedResults, SpecificExpectedResults]] = Seq(
@@ -145,14 +153,14 @@ class ReliefAtSourcePaymentsAndTaxReliefAmountControllerISpec extends Integratio
       import user.commonExpectedResults._
 
       s"language is ${welshTest(user.isWelsh)} and request is from an ${agentTest(user.isAgent)}" should {
-        "render Total payments into relief at source (RAS) pensions, plus basic rate tax relief page with no value when no cya data" which {
+        "render how much did you pay into your workplace pensions amount page with no pre filling" which {
           lazy val result: WSResponse = {
             dropPensionsDB()
             authoriseAgentOrIndividual(user.isAgent)
             val pensionsViewModel = aPaymentsIntoPensionViewModel.copy(
-              rasPensionPaymentQuestion = Some(true), totalRASPaymentsAndTaxRelief = None)
+              workplacePensionPaymentsQuestion = Some(true), totalWorkplacePensionPayments = None)
             insertCyaData(pensionsUsersData(isPrior = false, aPensionsCYAModel.copy(paymentsIntoPension = pensionsViewModel)), aUserRequest)
-            urlGet(fullUrl(reliefAtSourcePaymentsAndTaxReliefAmountUrl(taxYearEOY)),
+            urlGet(fullUrl(workplacePensionAmount(taxYearEOY)),
               user.isWelsh, follow = false, headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
           }
 
@@ -162,31 +170,31 @@ class ReliefAtSourcePaymentsAndTaxReliefAmountControllerISpec extends Integratio
             result.status shouldBe OK
           }
 
-          titleCheck(expectedTitle)
-          h1Check(expectedHeading)
+          titleCheck(user.specificExpectedResults.get.expectedTitle)
+          h1Check(user.specificExpectedResults.get.expectedHeading)
           captionCheck(expectedCaption(taxYearEOY), captionSelector)
-          textOnPageCheck(user.specificExpectedResults.get.expectedWhereToFind, paragraphSelector(2))
-          textOnPageCheck(user.specificExpectedResults.get.expectedHowToWorkOut, paragraphSelector(3))
-          textOnPageCheck(expectedCalculationHeading, insetSpanText(1))
-          textOnPageCheck(expectedExampleCalculation, insetSpanText(2))
+          textOnPageCheck(expectedParagraph, paragraphSelector(2))
+          textOnPageCheck(user.specificExpectedResults.get.expectedBullet1, bulletListSelector(1))
+          textOnPageCheck(user.specificExpectedResults.get.expectedBullet2, bulletListSelector(2))
+          textOnPageCheck(user.specificExpectedResults.get.expectedYouCanFindThisOut, paragraphSelector(4))
           textOnPageCheck(hintText, hintTextSelector)
           textOnPageCheck(poundPrefixText, poundPrefixSelector)
           inputFieldValueCheck(amountInputName, inputSelector, "")
           buttonCheck(buttonText, continueButtonSelector)
-          formPostLinkCheck(reliefAtSourcePaymentsAndTaxReliefAmountUrl(taxYearEOY), formSelector)
+          formPostLinkCheck(workplacePensionAmount(taxYearEOY), formSelector)
           welshToggleCheck(user.isWelsh)
         }
 
-        "render Total payments into relief at source (RAS) pensions, plus basic rate tax relief page prefilled when cya data" which {
+        "render how much did you pay into your workplace pensions amount page when cya data" which {
 
           val existingAmount: String = "999.88"
           lazy val result: WSResponse = {
             dropPensionsDB()
             authoriseAgentOrIndividual(user.isAgent)
             val pensionsViewModel = aPaymentsIntoPensionViewModel.copy(
-              rasPensionPaymentQuestion = Some(true), totalRASPaymentsAndTaxRelief = Some(BigDecimal(existingAmount)))
+              workplacePensionPaymentsQuestion = Some(true), totalWorkplacePensionPayments = Some(BigDecimal(existingAmount)))
             insertCyaData(pensionsUsersData(isPrior = false, aPensionsCYAModel.copy(paymentsIntoPension = pensionsViewModel)), aUserRequest)
-            urlGet(fullUrl(reliefAtSourcePaymentsAndTaxReliefAmountUrl(taxYearEOY)), user.isWelsh, follow = false,
+            urlGet(fullUrl(workplacePensionAmount(taxYearEOY)), user.isWelsh, follow = false,
               headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
           }
 
@@ -195,59 +203,56 @@ class ReliefAtSourcePaymentsAndTaxReliefAmountControllerISpec extends Integratio
           "has an OK status" in {
             result.status shouldBe OK
           }
-
-          titleCheck(expectedTitle)
-          h1Check(expectedHeading)
+          titleCheck(user.specificExpectedResults.get.expectedTitle)
+          h1Check(user.specificExpectedResults.get.expectedHeading)
           captionCheck(expectedCaption(taxYearEOY), captionSelector)
-          textOnPageCheck(user.specificExpectedResults.get.expectedWhereToFind, paragraphSelector(2))
-          textOnPageCheck(user.specificExpectedResults.get.expectedHowToWorkOut, paragraphSelector(3))
-          textOnPageCheck(expectedCalculationHeading, insetSpanText(1))
-          textOnPageCheck(expectedExampleCalculation, insetSpanText(2))
+          textOnPageCheck(expectedParagraph, paragraphSelector(2))
+          textOnPageCheck(user.specificExpectedResults.get.expectedBullet1, bulletListSelector(1))
+          textOnPageCheck(user.specificExpectedResults.get.expectedBullet2, bulletListSelector(2))
+          textOnPageCheck(user.specificExpectedResults.get.expectedYouCanFindThisOut, paragraphSelector(4))
           textOnPageCheck(hintText, hintTextSelector)
           textOnPageCheck(poundPrefixText, poundPrefixSelector)
           inputFieldValueCheck(amountInputName, inputSelector, existingAmount)
           buttonCheck(buttonText, continueButtonSelector)
-          formPostLinkCheck(reliefAtSourcePaymentsAndTaxReliefAmountUrl(taxYearEOY), formSelector)
+          formPostLinkCheck(workplacePensionAmount(taxYearEOY), formSelector)
           welshToggleCheck(user.isWelsh)
-
         }
-
       }
     }
-
-    "redirect to the rasPensionPaymentQuestion page if the question has not been answered" which {
+    "redirect to the Workplace pension question page if the question has not been answered" which {
       lazy val result: WSResponse = {
         dropPensionsDB()
         authoriseAgentOrIndividual(isAgent = false)
         val pensionsViewModel = aPaymentsIntoPensionViewModel.copy(
-          rasPensionPaymentQuestion = None, totalRASPaymentsAndTaxRelief = None)
+          workplacePensionPaymentsQuestion = None, totalWorkplacePensionPayments = None)
         insertCyaData(pensionsUsersData(isPrior = false, aPensionsCYAModel.copy(paymentsIntoPension = pensionsViewModel)), aUserRequest)
-        urlGet(fullUrl(reliefAtSourcePaymentsAndTaxReliefAmountUrl(taxYearEOY)), follow = false,
+        urlGet(fullUrl(workplacePensionAmount(taxYearEOY)), follow = false,
           headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
       }
 
       "has an SEE_OTHER status" in {
         result.status shouldBe SEE_OTHER
-        result.header("location").contains(reliefAtSourcePensionsUrl(taxYearEOY)) shouldBe true
+        result.header("location").contains(workplacePensionUrl(taxYearEOY)) shouldBe true
       }
 
     }
 
-    "redirect to the rasPensionPaymentQuestion page if the rasPensionPaymentQuestion question has been answered as false" which {
+    "redirect to the workplace question page if the workplaceQuestion has been answered as false" which {
       lazy val result: WSResponse = {
         dropPensionsDB()
         authoriseAgentOrIndividual(isAgent = false)
         val pensionsViewModel = aPaymentsIntoPensionViewModel.copy(
-          rasPensionPaymentQuestion = Some(false), totalRASPaymentsAndTaxRelief = None)
+          workplacePensionPaymentsQuestion = Some(false), totalWorkplacePensionPayments = None)
         insertCyaData(pensionsUsersData(isPrior = false, aPensionsCYAModel.copy(paymentsIntoPension = pensionsViewModel)), aUserRequest)
-        urlGet(fullUrl(reliefAtSourcePaymentsAndTaxReliefAmountUrl(taxYearEOY)), follow = false,
+        urlGet(fullUrl(workplacePensionAmount(taxYearEOY)), follow = false,
           headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
       }
 
       "has an SEE_OTHER status" in {
         result.status shouldBe SEE_OTHER
-        result.header("location").contains(reliefAtSourcePensionsUrl(taxYearEOY)) shouldBe true
+        result.header("location").contains(workplacePensionUrl(taxYearEOY)) shouldBe true
       }
+
     }
 
     "redirect to the CYA page if there is no session data" which {
@@ -255,7 +260,7 @@ class ReliefAtSourcePaymentsAndTaxReliefAmountControllerISpec extends Integratio
         dropPensionsDB()
         authoriseAgentOrIndividual(isAgent = false)
         // no cya insert
-        urlGet(fullUrl(reliefAtSourcePaymentsAndTaxReliefAmountUrl(taxYearEOY)), follow = false,
+        urlGet(fullUrl(workplacePensionAmount(taxYearEOY)), follow = false,
           headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
       }
 
@@ -265,7 +270,6 @@ class ReliefAtSourcePaymentsAndTaxReliefAmountControllerISpec extends Integratio
       }
 
     }
-
   }
 
   ".submit" should {
@@ -284,10 +288,10 @@ class ReliefAtSourcePaymentsAndTaxReliefAmountControllerISpec extends Integratio
           lazy val result: WSResponse = {
             dropPensionsDB()
             val pensionsViewModel = aPaymentsIntoPensionViewModel.copy(
-              rasPensionPaymentQuestion = Some(true), totalRASPaymentsAndTaxRelief = None)
+              workplacePensionPaymentsQuestion = Some(true), totalWorkplacePensionPayments = None)
             insertCyaData(pensionsUsersData(isPrior = false, aPensionsCYAModel.copy(paymentsIntoPension = pensionsViewModel)), aUserRequest)
             authoriseAgentOrIndividual(user.isAgent)
-            urlPost(fullUrl(reliefAtSourcePaymentsAndTaxReliefAmountUrl(taxYearEOY)), body = emptyForm, welsh = user.isWelsh,
+            urlPost(fullUrl(workplacePensionAmount(taxYearEOY)), body = emptyForm, welsh = user.isWelsh,
               follow = false, headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
           }
 
@@ -297,18 +301,18 @@ class ReliefAtSourcePaymentsAndTaxReliefAmountControllerISpec extends Integratio
 
           implicit def document: () => Document = () => Jsoup.parse(result.body)
 
-          titleCheck(expectedErrorTitle)
-          h1Check(expectedHeading)
+          titleCheck(user.specificExpectedResults.get.expectedErrorTitle)
+          h1Check(user.specificExpectedResults.get.expectedHeading)
           captionCheck(expectedCaption(taxYearEOY), captionSelector)
-          textOnPageCheck(user.specificExpectedResults.get.expectedWhereToFind, paragraphSelector(2))
-          textOnPageCheck(user.specificExpectedResults.get.expectedHowToWorkOut, paragraphSelector(3))
-          textOnPageCheck(expectedCalculationHeading, insetSpanText(1))
-          textOnPageCheck(expectedExampleCalculation, insetSpanText(2))
+          textOnPageCheck(expectedParagraph, paragraphSelector(2))
+          textOnPageCheck(user.specificExpectedResults.get.expectedBullet1, bulletListSelector(1))
+          textOnPageCheck(user.specificExpectedResults.get.expectedBullet2, bulletListSelector(2))
+          textOnPageCheck(user.specificExpectedResults.get.expectedYouCanFindThisOut, paragraphSelector(4))
           textOnPageCheck(hintText, hintTextSelector)
           textOnPageCheck(poundPrefixText, poundPrefixSelector)
           inputFieldValueCheck(amountInputName, inputSelector, amountEmpty)
           buttonCheck(buttonText, continueButtonSelector)
-          formPostLinkCheck(reliefAtSourcePaymentsAndTaxReliefAmountUrl(taxYearEOY), formSelector)
+          formPostLinkCheck(workplacePensionAmount(taxYearEOY), formSelector)
           errorSummaryCheck(emptyErrorText, expectedErrorHref)
           errorAboveElementCheck(emptyErrorText)
           welshToggleCheck(user.isWelsh)
@@ -322,10 +326,10 @@ class ReliefAtSourcePaymentsAndTaxReliefAmountControllerISpec extends Integratio
           lazy val result: WSResponse = {
             dropPensionsDB()
             val pensionsViewModel = aPaymentsIntoPensionViewModel.copy(
-              rasPensionPaymentQuestion = Some(true), totalRASPaymentsAndTaxRelief = None)
+              workplacePensionPaymentsQuestion = Some(true), totalWorkplacePensionPayments = None)
             insertCyaData(pensionsUsersData(isPrior = false, aPensionsCYAModel.copy(paymentsIntoPension = pensionsViewModel)), aUserRequest)
             authoriseAgentOrIndividual(user.isAgent)
-            urlPost(fullUrl(reliefAtSourcePaymentsAndTaxReliefAmountUrl(taxYearEOY)), body = invalidFormatForm, welsh = user.isWelsh,
+            urlPost(fullUrl(workplacePensionAmount(taxYearEOY)), body = invalidFormatForm, welsh = user.isWelsh,
               follow = false, headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
           }
 
@@ -335,18 +339,18 @@ class ReliefAtSourcePaymentsAndTaxReliefAmountControllerISpec extends Integratio
 
           implicit def document: () => Document = () => Jsoup.parse(result.body)
 
-          titleCheck(expectedErrorTitle)
-          h1Check(expectedHeading)
+          titleCheck(user.specificExpectedResults.get.expectedErrorTitle)
+          h1Check(user.specificExpectedResults.get.expectedHeading)
           captionCheck(expectedCaption(taxYearEOY), captionSelector)
-          textOnPageCheck(user.specificExpectedResults.get.expectedWhereToFind, paragraphSelector(2))
-          textOnPageCheck(user.specificExpectedResults.get.expectedHowToWorkOut, paragraphSelector(3))
-          textOnPageCheck(expectedCalculationHeading, insetSpanText(1))
-          textOnPageCheck(expectedExampleCalculation, insetSpanText(2))
+          textOnPageCheck(expectedParagraph, paragraphSelector(2))
+          textOnPageCheck(user.specificExpectedResults.get.expectedBullet1, bulletListSelector(1))
+          textOnPageCheck(user.specificExpectedResults.get.expectedBullet2, bulletListSelector(2))
+          textOnPageCheck(user.specificExpectedResults.get.expectedYouCanFindThisOut, paragraphSelector(4))
           textOnPageCheck(hintText, hintTextSelector)
           textOnPageCheck(poundPrefixText, poundPrefixSelector)
           inputFieldValueCheck(amountInputName, inputSelector, amountInvalidFormat)
           buttonCheck(buttonText, continueButtonSelector)
-          formPostLinkCheck(reliefAtSourcePaymentsAndTaxReliefAmountUrl(taxYearEOY), formSelector)
+          formPostLinkCheck(workplacePensionAmount(taxYearEOY), formSelector)
           errorSummaryCheck(invalidFormatErrorText, expectedErrorHref)
           errorAboveElementCheck(invalidFormatErrorText)
           welshToggleCheck(user.isWelsh)
@@ -360,10 +364,10 @@ class ReliefAtSourcePaymentsAndTaxReliefAmountControllerISpec extends Integratio
           lazy val result: WSResponse = {
             dropPensionsDB()
             val pensionsViewModel = aPaymentsIntoPensionViewModel.copy(
-              rasPensionPaymentQuestion = Some(true), totalRASPaymentsAndTaxRelief = None)
+              workplacePensionPaymentsQuestion = Some(true), totalWorkplacePensionPayments = None)
             insertCyaData(pensionsUsersData(isPrior = false, aPensionsCYAModel.copy(paymentsIntoPension = pensionsViewModel)), aUserRequest)
             authoriseAgentOrIndividual(user.isAgent)
-            urlPost(fullUrl(reliefAtSourcePaymentsAndTaxReliefAmountUrl(taxYearEOY)), body = overMaximumForm, welsh = user.isWelsh,
+            urlPost(fullUrl(workplacePensionAmount(taxYearEOY)), body = overMaximumForm, welsh = user.isWelsh,
               follow = false, headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
           }
 
@@ -373,27 +377,24 @@ class ReliefAtSourcePaymentsAndTaxReliefAmountControllerISpec extends Integratio
 
           implicit def document: () => Document = () => Jsoup.parse(result.body)
 
-          titleCheck(expectedErrorTitle)
-          h1Check(expectedHeading)
+          titleCheck(user.specificExpectedResults.get.expectedErrorTitle)
+          h1Check(user.specificExpectedResults.get.expectedHeading)
           captionCheck(expectedCaption(taxYearEOY), captionSelector)
-          textOnPageCheck(user.specificExpectedResults.get.expectedWhereToFind, paragraphSelector(2))
-          textOnPageCheck(user.specificExpectedResults.get.expectedHowToWorkOut, paragraphSelector(3))
-          textOnPageCheck(expectedCalculationHeading, insetSpanText(1))
-          textOnPageCheck(expectedExampleCalculation, insetSpanText(2))
+          textOnPageCheck(expectedParagraph, paragraphSelector(2))
+          textOnPageCheck(user.specificExpectedResults.get.expectedBullet1, bulletListSelector(1))
+          textOnPageCheck(user.specificExpectedResults.get.expectedBullet2, bulletListSelector(2))
+          textOnPageCheck(user.specificExpectedResults.get.expectedYouCanFindThisOut, paragraphSelector(4))
           textOnPageCheck(hintText, hintTextSelector)
           textOnPageCheck(poundPrefixText, poundPrefixSelector)
           inputFieldValueCheck(amountInputName, inputSelector, amountOverMaximum)
           buttonCheck(buttonText, continueButtonSelector)
-          formPostLinkCheck(reliefAtSourcePaymentsAndTaxReliefAmountUrl(taxYearEOY), formSelector)
+          formPostLinkCheck(workplacePensionAmount(taxYearEOY), formSelector)
           errorSummaryCheck(maxAmountErrorText, expectedErrorHref)
           errorAboveElementCheck(maxAmountErrorText)
           welshToggleCheck(user.isWelsh)
         }
-
       }
-
     }
-
     "redirect to the correct page when a valid amount is submitted and update the session amount" which {
 
       val validAmount = "100.22"
@@ -402,23 +403,24 @@ class ReliefAtSourcePaymentsAndTaxReliefAmountControllerISpec extends Integratio
       lazy val result: WSResponse = {
         dropPensionsDB()
         val pensionsViewModel = aPaymentsIntoPensionViewModel.copy(
-          rasPensionPaymentQuestion = Some(true), totalRASPaymentsAndTaxRelief = None)
+          workplacePensionPaymentsQuestion = Some(true), totalWorkplacePensionPayments = None)
         insertCyaData(pensionsUsersData(isPrior = false, aPensionsCYAModel.copy(paymentsIntoPension = pensionsViewModel)), aUserRequest)
         authoriseAgentOrIndividual(isAgent = false)
-        urlPost(fullUrl(reliefAtSourcePaymentsAndTaxReliefAmountUrl(taxYearEOY)), body = validForm, follow = false,
+        urlPost(fullUrl(workplacePensionAmount(taxYearEOY)), body = validForm, follow = false,
           headers = Seq(HeaderNames.COOKIE -> playSessionCookies(taxYearEOY)))
       }
 
       "has a SEE_OTHER(303) status" in {
         result.status shouldBe SEE_OTHER
-        result.header("location") shouldBe Some(reliefAtSourceOneOffPaymentsUrl(taxYearEOY))
+        result.header("location") shouldBe Some(checkPaymentsIntoPensionCyaUrl(taxYearEOY))
       }
 
-      "updates retirement annuity contract payments question to Some(true)" in {
+      "updates workplace amount" in {
         lazy val cyaModel = findCyaData(taxYearEOY, aUserRequest).get
-        cyaModel.pensions.paymentsIntoPension.totalRASPaymentsAndTaxRelief shouldBe Some(BigDecimal(validAmount))
+        cyaModel.pensions.paymentsIntoPension.totalWorkplacePensionPayments shouldBe Some(BigDecimal(validAmount))
       }
     }
 
   }
 }
+// scalastyle:on magic.number
